Better implementation for config commands.
